+++ packages/components/ContextualMenu/src/ContextualMenuItem.tsx
@@ -15,7 +15,7 @@ import { Text } from '@fluentui-react-native/text';
 import { settings } from './ContextualMenuItem.settings';
 import { backgroundColorTokens, borderTokens, textTokens, foregroundColorTokens, getPaletteFromTheme } from '@fluentui-react-native/tokens';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
-import { useAsPressable, useKeyCallback, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
+import { useAsPressable, useKeyUpProps, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
 import { CMContext } from './ContextualMenu';
 import { Icon } from '@fluentui-react-native/icon';
 import { createIconProps } from '@fluentui-react-native/interactive-hooks';
@@ -63,7 +63,7 @@ export const ContextualMenuItem = compose<ContextualMenuItemType>({
 
     const pressable = useAsPressable({ ...rest, onPress: onItemClick, onHoverIn: onItemHoverIn });
 
-    const onKeyUp = useKeyCallback(onItemClick, ' ', 'Enter');
+    const onKeyUpProps = useKeyUpProps(onItemClick, ' ', 'Enter');
 
     // set up state
     const state: ContextualMenuItemState = {
@@ -103,7 +103,7 @@ export const ContextualMenuItem = compose<ContextualMenuItemType>({
       root: {
         ...pressable.props,
         ref: cmRef,
-        onKeyUp,
+        ...onKeyUpProps,
         onMouseEnter,
         onMouseLeave,
         accessible: true,