+++ packages/experimental/Button/src/Button.styling.ts
@@ -1,9 +1,9 @@
-import { buttonName, ButtonCoreTokens, ButtonTokens, ButtonSlotProps, ButtonProps, ButtonSize } from './Button.types';
+import { buttonName, ButtonCoreTokens, ButtonTokens, ButtonSlotProps, ButtonPropsWithInnerRef, ButtonSize } from './Button.types';
 import { Theme, UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
-import { borderStyles, layoutStyles, fontStyles, shadowStyles } from '@fluentui-react-native/tokens';
+import { borderStyles, layoutStyles, fontStyles, shadowStyles, FontTokens } from '@fluentui-react-native/tokens';
 import { defaultButtonTokens } from './ButtonTokens';
 import { defaultButtonColorTokens } from './ButtonColorTokens';
-import { Platform } from 'react-native';
+import { Platform, ColorValue } from 'react-native';
 import { getTextMarginAdjustment } from '@fluentui-react-native/styling-utils';
 
 export const buttonCoreStates: (keyof ButtonCoreTokens)[] = ['hovered', 'focused', 'pressed', 'disabled', 'hasContent', 'hasIcon'];
@@ -13,7 +13,6 @@ export const buttonStates: (keyof ButtonTokens)[] = [
   'primary',
   'subtle',
   'hovered',
-  'focused',
   'pressed',
   'disabled',
   'small',
@@ -24,9 +23,10 @@ export const buttonStates: (keyof ButtonTokens)[] = [
   'rounded',
   'circular',
   'square',
+  'focused',
 ];
 
-export const stylingSettings: UseStylingOptions<ButtonProps, ButtonSlotProps, ButtonTokens> = {
+export const stylingSettings: UseStylingOptions<ButtonPropsWithInnerRef, ButtonSlotProps, ButtonTokens> = {
   tokens: [defaultButtonTokens, defaultButtonColorTokens, buttonName],
   states: buttonStates,
   slotProps: {
@@ -49,18 +49,9 @@ export const stylingSettings: UseStylingOptions<ButtonProps, ButtonSlotProps, Bu
     ),
     content: buildProps(
       (tokens: ButtonTokens, theme: Theme) => {
-        const spacingIconContent = tokens.spacingIconContent
-          ? {
-              marginLeft: tokens.spacingIconContent,
-              marginRight: tokens.spacingIconContent,
-            }
-          : {};
         return {
           style: {
-            color: tokens.color,
-            ...getTextMarginAdjustment(),
-            ...spacingIconContent,
-            ...fontStyles.from(tokens, theme),
+            ...contentStyling(tokens, theme, tokens.color, tokens),
           },
         };
       },
@@ -87,4 +78,19 @@ export const getDefaultSize = (): ButtonSize => {
   }
 
   return 'medium';
+};
+
+export const contentStyling = (tokens: ButtonTokens, theme: Theme, contentColor: ColorValue, fontStylesTokens: FontTokens) => {
+  const spacingIconContent = tokens.spacingIconContent
+    ? {
+        marginLeft: tokens.spacingIconContent,
+        marginRight: tokens.spacingIconContent,
+      }
+    : {};
+  return {
+    color: contentColor,
+    ...getTextMarginAdjustment(),
+    ...spacingIconContent,
+    ...fontStyles.from(fontStylesTokens, theme),
+  };
 };