+++ packages/experimental/Button/src/Button.tsx
@@ -1,13 +1,14 @@
 /** @jsx withSlots */
 import * as React from 'react';
-import { View, ActivityIndicator } from 'react-native';
-import { buttonName, ButtonType, ButtonProps } from './Button.types';
+import { View } from 'react-native';
+import { ActivityIndicator } from '@fluentui-react-native/experimental-activity-indicator';
+import { buttonName, ButtonType, ButtonProps, ButtonPropsWithInnerRef } from './Button.types';
 import { Text } from '@fluentui-react-native/experimental-text';
 import { stylingSettings, getDefaultSize } from './Button.styling';
 import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
 import { useButton } from './useButton';
 import { Icon } from '@fluentui-react-native/icon';
-import { createIconProps, IPressableState } from '@fluentui-react-native/interactive-hooks';
+import { createIconProps, IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
 
 /**
  * A function which determines if a set of styles should be applied to the compoent given the current state and props of the button.
@@ -17,7 +18,7 @@ import { createIconProps, IPressableState } from '@fluentui-react-native/interac
  * @param userProps The props that were passed into the button
  * @returns Whether the styles that are assigned to the layer should be applied to the button
  */
-export const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonProps): boolean => {
+export const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonPropsWithInnerRef): boolean => {
   return (
     state[layer] ||
     userProps[layer] ||
@@ -31,7 +32,7 @@ export const buttonLookup = (layer: string, state: IPressableState, userProps: B
   );
 };
 
-export const Button = compose<ButtonType>({
+const ButtonComposed = compose<ButtonType>({
   displayName: buttonName,
   ...stylingSettings,
   slots: {
@@ -39,17 +40,27 @@ export const Button = compose<ButtonType>({
     icon: Icon,
     content: Text,
   },
-  render: (userProps: ButtonProps, useSlots: UseSlots<ButtonType>) => {
+  render: (userProps: ButtonPropsWithInnerRef, useSlots: UseSlots<ButtonType>) => {
     const button = useButton(userProps);
     const iconProps = createIconProps(userProps.icon);
     // grab the styled slots
     const Slots = useSlots(userProps, (layer) => buttonLookup(layer, button.state, userProps));
 
     // now return the handler for finishing render
-    return (final: ButtonProps, ...children: React.ReactNode[]) => {
-      const { icon, iconPosition, loading, ...mergedProps } = mergeProps(button.props, final);
+    return (final: ButtonPropsWithInnerRef, children: React.ReactNode[]) => {
+      const { icon, iconPosition, loading, accessiblityLabel, ...mergedProps } = mergeProps(button.props, final);
+      let childText = '';
+      if (accessiblityLabel === undefined) {
+        React.Children.forEach(children, (child) => {
+          if (typeof child === 'string') {
+            childText = child; // We only automatically support the one child string.
+          }
+        });
+      }
+      const label = accessiblityLabel ? accessiblityLabel : childText;
+
       return (
-        <Slots.root {...mergedProps}>
+        <Slots.root {...mergedProps} accessibilityLabel={label}>
           {icon && iconPosition === 'before' && <Slots.icon {...iconProps} />}
           {loading && <ActivityIndicator />}
           {React.Children.map(children, (child) =>
@@ -61,5 +72,7 @@ export const Button = compose<ButtonType>({
     };
   },
 });
+
+export const Button = React.forwardRef<IFocusable, ButtonProps>((props, ref) => <ButtonComposed {...props} innerRef={ref} />);
 
 export default Button;