+++ packages/experimental/Button/src/FAB/FABCore.tsx
@@ -7,8 +7,8 @@ import { stylingSettings } from './FAB.styling';
 import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
 import { useButton } from '../useButton';
 import { Icon } from '@fluentui-react-native/icon';
-import { createIconProps, IPressableState } from '@fluentui-react-native/interactive-hooks';
-import { ButtonCoreProps } from '../Button.types';
+import { createIconProps, IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
+import { ButtonCorePropsWithInnerRef, ButtonCoreProps } from '../Button.types';
 
 /**
  * A function which determines if a set of styles should be applied to the compoent given the current state and props of the button.
@@ -18,11 +18,11 @@ import { ButtonCoreProps } from '../Button.types';
  * @param userProps The props that were passed into the button
  * @returns Whether the styles that are assigned to the layer should be applied to the button
  */
-const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonCoreProps): boolean => {
-  return state[layer] || userProps[layer] || (layer === 'hasContent' && userProps.content) || (layer === 'hasIcon' && userProps.icon);
+const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonCorePropsWithInnerRef): boolean => {
+  return state[layer] || userProps[layer] || (layer === 'hasContent' && !userProps.iconOnly) || (layer === 'hasIcon' && userProps.icon);
 };
 
-export const FAB = compose<FABType>({
+const FABComposed = compose<FABType>({
   displayName: fabName,
   ...stylingSettings,
   slots: {
@@ -30,8 +30,8 @@ export const FAB = compose<FABType>({
     icon: Icon,
     content: Text,
   },
-  render: (userProps: ButtonCoreProps, useSlots: UseSlots<FABType>) => {
-    const { icon, content, onClick, ...rest } = userProps;
+  render: (userProps: ButtonCorePropsWithInnerRef, useSlots: UseSlots<FABType>) => {
+    const { icon, onClick, ...rest } = userProps;
     const iconProps = createIconProps(userProps.icon);
 
     const button = useButton(rest);
@@ -40,18 +40,21 @@ export const FAB = compose<FABType>({
     const Slots = useSlots(userProps, (layer) => buttonLookup(layer, button.state, userProps));
 
     // now return the handler for finishing render
-    return (final: ButtonCoreProps, ...children: React.ReactNode[]) => {
+    return (final: ButtonCorePropsWithInnerRef, ...children: React.ReactNode[]) => {
       const mergedProps = mergeProps(button.props, final);
 
       return (
         <Slots.root {...mergedProps}>
           {icon && <Slots.icon {...iconProps} />}
-          {content && <Slots.content key="content">{content}</Slots.content>}
-          {children}
+          {React.Children.map(children, (child) =>
+            typeof child === 'string' ? <Slots.content key="content">{child}</Slots.content> : child,
+          )}
         </Slots.root>
       );
     };
   },
 });
+
+export const FAB = React.forwardRef<IFocusable, ButtonCoreProps>((props, ref) => <FABComposed {...props} innerRef={ref} />);
 
 export default FAB;