+++ packages/experimental/Button/src/ToggleButton/ToggleButton.test.tsx
@@ -3,6 +3,6 @@ import { ToggleButton } from './ToggleButton';
 import * as renderer from 'react-test-renderer';
 
 it('ToggleButton default', () => {
-  const tree = renderer.create(<ToggleButton content="Default Button" />).toJSON();
+  const tree = renderer.create(<ToggleButton>Default Button</ToggleButton>).toJSON();
   expect(tree).toMatchSnapshot();
 });