+++ packages/theming/theme-tokens/src/dark-macos/tokens-alias.ts
@@ -1,3 +1,3 @@
-import aliasTokens from './reactnative/tokens-aliases.json';
+import aliasTokens from '../generated/dark-macos/reactnative/tokens-aliases.json';
 
 export default aliasTokens;