+++ packages/deprecated/foundation-compose/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @uifabricshared/foundation-compose
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 01:26:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.11.2
+
+Fri, 17 Dec 2021 01:26:42 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.2
+
+## 1.11.1
+
+Thu, 16 Dec 2021 19:10:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.1
 
 ## 1.11.0