+++ packages/experimental/Tabs/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui-react-native/experimental-tabs
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 19:53:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.3
+
+Fri, 17 Dec 2021 19:53:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.3
+- Bump @fluentui-react-native/interactive-hooks to v0.11.3
+- Bump @fluentui-react-native/focus-zone to v0.8.3
+- Bump @fluentui-react-native/experimental-text to v0.7.3
+- Bump @fluentui-react-native/icon to v0.10.2
+
+## 0.4.2
+
+Fri, 17 Dec 2021 01:26:42 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.2
+- Bump @fluentui-react-native/interactive-hooks to v0.11.2
+- Bump @fluentui-react-native/focus-zone to v0.8.2
+- Bump @fluentui-react-native/experimental-text to v0.7.2
+- Bump @fluentui-react-native/icon to v0.10.1
+
+## 0.4.1
+
+Thu, 16 Dec 2021 19:10:27 GMT
+
+### Patches
+
+- Refactor useKeyCallback (sanajmi@microsoft.com)
+- Rename Jest snapshots to be platform agnostic  (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.7.1
+- Bump @fluentui-react-native/interactive-hooks to v0.11.1
+- Bump @fluentui-react-native/focus-zone to v0.8.1
+- Bump @fluentui-react-native/experimental-text to v0.7.1
+- Bump @fluentui-react-native/tokens to v0.11.1
+- Bump @fluentui-react-native/icon to v0.10.0
+- Bump @fluentui-react-native/adapters to v0.8.1
 
 ## 0.4.0