+++ packages/framework/use-slots/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/use-slots
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 19:53:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.1
+
+Fri, 17 Dec 2021 19:53:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/use-slot to v0.2.1
 
 ## 0.6.0