+++ packages/utils/adapters/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/adapters
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Thu, 16 Dec 2021 19:10:30 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.1
+
+Thu, 16 Dec 2021 19:10:30 GMT
+
+### Patches
+
+- Have web adapter export React Native core types (sanajmi@microsoft.com)
+- Fix build (ruaraki@microsoft.com)
 
 ## 0.8.0